requirements.txt
numpy==1.17.3
opencv-python==4.1.1.26
pyserial==3.4
